Similarity and classLDA_full
Pushing updates to the class-based LDA for the documents, this put their class membership in the database.

I am also adding code "similarity.py" which is a script that takes a user query, and provides semantic search across our corpus using the least cosine distance to the BERT embeddings. I'll add comment to the code later to explain it.